+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 * Include current row with currentSelectedRows list in case of grid selection mode is row.
 * add missing arguments to gridConfiguration.copyWith
 * Update column menu example allow merging with the default menu
-
+* Simplify the column menu delegate, adding example for removing one of the default menu items
 
 ## [1.6.2] - 2025. 03. 19